+++ GeoClone/HowMenu.cpp
@@ -39,19 +39,19 @@ void HowMenu::update(std::shared_ptr<Draw> &renderer){
 	renderer->speedRender(0, 0, _menuAssets[0]->getWidth(), _menuAssets[0]->getHeight(), _menuAssets[0]->getTexture());
 
 	//Render the controls
-	renderer->Render(_startX - 400, _startY - 380, _menuAssets[5]->getWidth(), _menuAssets[5]->getHeight(), _menuAssets[5]->getTexture());
+	renderer->Render(_startX - 400.f, _startY - 380.f, _menuAssets[5]->getWidth(), _menuAssets[5]->getHeight(), _menuAssets[5]->getTexture());
 
 	HAPI->ChangeFont("Tahoma", 30, 1000);
 	for(size_t curItem = 0; curItem < _menuItems.size(); curItem++){
 		if(curItem == _selectedItem){
-			HAPI->RenderText(_startX + 20, _startY + 140 + (curItem * 80), HAPI_TColour(180, 255, 0), _menuItems[curItem]);
+			HAPI->RenderText(_startX + 20, _startY + 140 + (static_cast<int>(curItem) * 80), HAPI_TColour(180, 255, 0), _menuItems[curItem]);
 			renderer->Render(_startX + 20 - (_menuAssets[1]->getWidth() / 2) + 80.f, 
-							_startY + 140  + (curItem * 80) - (_menuAssets[1]->getHeight() / 2) + 18.f,
+							_startY + 140  + (static_cast<int>(curItem) * 80) - (_menuAssets[1]->getHeight() / 2) + 18.f,
 							_menuAssets[1]->getWidth(), _menuAssets[1]->getHeight(), _menuAssets[1]->getTexture());
 		} else {
-			HAPI->RenderText(_startX, _startY + (curItem * 80), HAPI_TColour(0, 255, 0), _menuItems[curItem]);
+			HAPI->RenderText(_startX, _startY + (static_cast<int>(curItem) * 80), HAPI_TColour(0, 255, 0), _menuItems[curItem]);
 			renderer->Render(_startX - (_menuAssets[2]->getWidth() / 2) + 80.f, 
-							_startY + (curItem * 80) - (_menuAssets[2]->getHeight() / 2) + 18.f,
+							_startY + (static_cast<int>(curItem) * 80) - (_menuAssets[2]->getHeight() / 2) + 18.f,
 							_menuAssets[2]->getWidth(), _menuAssets[2]->getHeight(), _menuAssets[2]->getTexture());
 		}
 	}
@@ -71,7 +71,7 @@ void HowMenu::changeItem(int change){
 	_selectedItem += change;
 
 	if(_selectedItem < 0)
-		_selectedItem =  _menuItems.size() - 1;
+		_selectedItem =  static_cast<int>(_menuItems.size() - 1);
 
 	if(static_cast<size_t>(_selectedItem) > _menuItems.size() - 1){
 		_selectedItem = 0;
@@ -90,27 +90,27 @@ void HowMenu::selectedItem(bool &inMenues, std::shared_ptr<MenuState> &gameMenu,
 
 bool HowMenu::setupAssets(){
 	_menuAssets.push_back(std::shared_ptr<Asset>(new Asset()));
-	if(!(_menuAssets.back()->Load("images/select_menu_bg.png", LOAD_IMAGE, "menu_background", _menuAssets.size() - 1)))	//Borrow the select menu BG as it's pretty fitting for this menu.
+	if(!(_menuAssets.back()->Load("images/select_menu_bg.png", LOAD_IMAGE, "menu_background", static_cast<int>(_menuAssets.size() - 1))))	//Borrow the select menu BG as it's pretty fitting for this menu.
 		return false;
 
 	_menuAssets.push_back(std::shared_ptr<Asset>(new Asset()));
-	if(!(_menuAssets.back()->Load("images/main_menu_active.png", LOAD_IMAGE, "menu_active", _menuAssets.size() - 1)))
+	if(!(_menuAssets.back()->Load("images/main_menu_active.png", LOAD_IMAGE, "menu_active",  static_cast<int>(_menuAssets.size() - 1))))
 		return false;
 
 	_menuAssets.push_back(std::shared_ptr<Asset>(new Asset()));
-	if(!(_menuAssets.back()->Load("images/main_menu_nonactive.png", LOAD_IMAGE, "menu_nonactive", _menuAssets.size() - 1)))
+	if(!(_menuAssets.back()->Load("images/main_menu_nonactive.png", LOAD_IMAGE, "menu_nonactive",  static_cast<int>(_menuAssets.size() - 1))))
 		return false;
 
 	_menuAssets.push_back(std::shared_ptr<Asset>(new Asset()));
-	if(!(_menuAssets.back()->Load("images/nav_updown.png", LOAD_IMAGE, "nav_updown", _menuAssets.size() - 1)))
+	if(!(_menuAssets.back()->Load("images/nav_updown.png", LOAD_IMAGE, "nav_updown",  static_cast<int>(_menuAssets.size() - 1))))
 		return false;
 
 	_menuAssets.push_back(std::shared_ptr<Asset>(new Asset()));
-	if(!(_menuAssets.back()->Load("images/controller/facebutton_a.png", LOAD_IMAGE, "controller_a", _menuAssets.size() - 1)))
+	if(!(_menuAssets.back()->Load("images/controller/facebutton_a.png", LOAD_IMAGE, "controller_a",  static_cast<int>(_menuAssets.size() - 1))))
 		return false;
 	
 	_menuAssets.push_back(std::shared_ptr<Asset>(new Asset()));
-	if(!(_menuAssets.back()->Load("images/menu_help_control.png", LOAD_IMAGE, "help_controls", _menuAssets.size() - 1)))
+	if(!(_menuAssets.back()->Load("images/menu_help_control.png", LOAD_IMAGE, "help_controls",  static_cast<int>(_menuAssets.size() - 1))))
 		return false;
 
 	return true;
